feat: 리덕스 도입
