+++ todayCody/frontend/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.5",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -11,6 +12,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "react-quill": "^2.0.0",
+    "react-redux": "^8.1.1",
     "react-router-dom": "^6.13.0",
     "react-scripts": "^5.0.1",
     "sass": "^1.62.1",
