+++ todayCody/frontend/src/store/configureStore.js
@@ -1,0 +1,6 @@
+import { configureStore } from "@reduxjs/toolkit";
+import sidebarReducer from "./sidebar";
+
+export const store = configureStore({
+  reducer: { sidebar: sidebarReducer },
+});
