+++ todayCody/frontend/src/store/sidebar.js
@@ -1,0 +1,19 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+const initialState = {
+  open: true,
+};
+
+export const sidebarSlice = createSlice({
+  name: "sidebar",
+  initialState,
+  reducers: {
+    toggleOpen: (state) => {
+      state.open = !state.open;
+    },
+  },
+});
+
+export const { toggleOpen } = sidebarSlice.actions;
+
+export default sidebarSlice.reducer;
